Mobile Block Screen (Temporarily Incompatibility)
For now, all mobile devices are not compatible with the software until we can develop the custom mobile CSS. So we give them this warning. Later on I'll link it to a page where you can learn more abuot the software and the team, how you can support it, where it's headed, and so forth.