Import/Export Presets - Preset Browser
Yea for the most part just added the ability to export and import presets. And also I made a preset browser. Right now it's pretty empty but I'm going to be loading that thang up real soon with some really fun visualizers.

And I hope as more of a community kind of comes along, the ability to share presets and what not will be really cool.

Maybe even spark weekly challenges where we host a livestream for the website, and we look through all of the visualizers.

Anyways this update is pretty nice, just upgrades the webapp as a whole.