Custom Restore Default
This was a fun little update to do. I tried to get the restore default to work before with an overcomplicated system, but I realized the easiest way to do it is to literally just export the default mode as a preset, keep it at the root of my presets folder, and then just call it with the button like anything else.

The really fun addition to this push is the custom panel and pop up that I made for when confirming if you want to restore the default preset. It's just a pixelized wooden panel png and a couple buttons, but it looks much nicer than just the boring stock chrome one.

Anyways that's it, solid update