Tweaked CD Collection CSS | Export Video & Screenshot Functionality
So there are two fun things that are being added in this push. For one, I spent a while tweaking a custom CSS for each CD in a seperate little folder so that I could have the perfect design. Once that was done, I was able to implement it and get rid of the placeholder template that gave it its look

Next, I wanted to finally add a way to capture video and scnreehost in the webapp without having to externally capture it.

I got it working for the most part, will definitely use some tweaks but its getting there. And I was able to use the scnreeshot feature to very quickly grab 8 screenshots for quick visualizer ideas that are not the new 8 cd covers. So that's fun, solid update