Sound Effects Galore
I added a few things in this update, but one of the main things that I'm so excited to keep diving more and more deep into is the sound engine that I was able to whip together.

Basically with what I was able to make, now you can assign a sound effect to literally any part of the website with a single line of code. And I plan on adding a ton of sound design, so that'll be really nice.

Anyways had a great time experimenting and making sounds, and so far it's turning out great