Sound FX & New Background Music
Yea I just went crazy with adding sound effects to all the buttons and what not. Now the website just feels so cozy and nostalgic with everything now, I love it.

And then just experimenting with puttings songs I've done before through a tight filter and bit crusher. I think eventually I'll make it where instead of printed FX like this, I'll keep the originals on the site and let them toggle the "Retro Filter" when they select the CD